+++ packages/dataparcels/.size-limit.json
@@ -20,15 +20,15 @@
         "path": "deleted.js"
     },
     {
-        "limit": "13 KB",
+        "limit": "12.3 KB",
         "path": "ParcelShape.js"
     },
     {
-        "limit": "5.75 KB",
+        "limit": "5.9 KB",
         "path": "ParcelNode.js"
     },
     {
-        "limit": "5.75 KB",
+        "limit": "6 KB",
         "path": "asNodes.js"
     },
     {
@@ -36,11 +36,11 @@
         "path": "asRaw.js"
     },
     {
-        "limit": "13 KB",
+        "limit": "12.3 KB",
         "path": "asShape.js"
     },
     {
-        "limit": "6 KB",
+        "limit": "5.3 KB",
         "path": "validation.js"
     }
 ]
