+++ packages/dataparcels/package.json
@@ -39,7 +39,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.1.5",
-    "unmutable": "^0.41.1"
+    "unmutable": "^0.46.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.1.2",
