--- packages/dataparcels/src/parcelData/move.js
@@ -1,3 +1,0 @@
-// @flow
-import swapOrMove from './swapOrMove';
-export default swapOrMove('move');
