--- packages/react-dataparcels-drag/.size-limit
@@ -1,6 +1,0 @@
-[
-    {
-        limit: "42 KB",
-        path: "lib/index.js"
-    }
-]
