+++ packages/react-dataparcels/package.json
@@ -48,7 +48,7 @@
     "@babel/runtime": "^7.1.5",
     "dataparcels": "^0.25.0-alpha.1",
     "is-promise": "^2.1.0",
-    "unmutable": "^0.41.1",
+    "unmutable": "^0.46.0",
     "use-debounce": "^1.1.3"
   },
   "devDependencies": {
